Sisulator added
The sisulator takes an XML file as input and converts this into a
JSON-compatible object according to a mapping ruleset. It will then
process a number of sisulets as specified in the given directive, which
recieve the object as input. The sisulets are parsed and the sisula
language substituted to JScript using regular expressions, after which
the JScript is evaluated and the output stored.