Polishing and More Added Functionality
Polished the example code.
RunAll.bat now supports UTF-8 environment variable values.
Added a trick that forces materialization in the split views.
However, the split logic is really slow when using SUBSTRING in
combination with CHARINDEX, which is why we started working on
substituting it for a CLR splitter.
The file finder now works optionally recursively.