Added functionality
Rowlength can now be specified in the <source> XML element to override
the Raw table [row] data type size, which otherwise defaults to 'max'.
Fixed bug where datafiletype in <source> was neglected.
Extended the assembly to include Try_Execute, which is used instead of
TRY_CAST in earlier versions of SQL Server (2008 and earlier).