Fixed bugs in BULK splitting
Working on getting the BULK splitting method runnable.
Fixed a bug where a non-existing table was being referenced in an update
statement after the BULK insert was completed.
Fixed a bug where inserts were not possible due to problems with the
table layout, so BULK now also has an insert view.