Bug fixes and improvements
Improved the XSLT transforms for conversion from the "old" framework to
this one.
Made all output from Sisulate.bat 7-bit ASCII for compatibility reasons.
The creation of format files is now more intelligent.
Removed an erroneous [row] reference in the insert view when bulk
splitting was used.
Made it possible to specify a subselect also for bulk splitted
configurations.
Empty calculations can now be created in order to get an empty column in
the typed table (that doesn't correspond to anything in the file).
Job deletion now first checks if the job exists, to avoid the harmless
error message that was shown previously.