Support for <map as="static">
Attributes can now be marked as="static", which disables the check exemplified by (source.Value <> target.AN_ATT_Attribute_Value) to trigger an UPDATE in the MERGE. This can be convenient if you want to silently disregard any future changes of an attribute value. Marking an attribute as="static" should therefore be done only after carefully considering that this is the desired behavior.